Move filter-methods to own module
